refactor: move jdbc options getter into JDBCConnectorConf
